objects ttl: extract ttl methods from index.go
